refactor: Improve error handling and tests
Some mixture of changes in, but all kinda depend on all other.

First, this PR replaces the []error slice with the compound error type `ParseErrors`. This give us two refactorings:
1) We can return an error instead of a slice.
2) We can use errors.As/errors.Is in the tests and check for that.
This refactoring already showed that the tests did not work on windows, as the tests failed with 'cannot read file'. This is due to the /home/ path, which is not working on Windows. With the error type checking, this bug is fixed.

The other refactor is that `EntryParserError` no longer has a factory function and takes the parent error directly. This also improves the error.As/error.Is handling.

Last but not least: use assert.NoError, assert.Empty, assert.Error, and assert.ErrorAs.