test(generate): Extend tests to assert generated DOT graph file content
Previously we only checked if files where created at all.
Now the main test also checks that a valid DOT file is created by unmarshalling it
again, and that the expected nodes and edges are written to the file.

As the simple directed graph we use is loaded without considering the node names/coordinates
we write to the file, the content is simply checked on string level, irgnoring the graph we
load when unmarshalling.
Creating a custom graph that can be unmarshalled using the coordinates as node IDs would be
possible, but effort outweighs the benefits for a simple testcase.